Refactor code and improve code readability
