+++ src/components/Loader/Loader.tsx
@@ -1,6 +1,7 @@
+import { FC } from 'react';
 import './Loader.scss';
 
-export const Loader = () => (
+export const Loader: FC = () => (
   <div className="Loader" data-cy="loader">
     <div className="Loader__content" />
   </div>
