+++ src/components/NavigationLink/NavigationLink.tsx
@@ -1,19 +1,19 @@
-import React from 'react';
+import { FC, ReactNode } from 'react';
+import { To, NavLink } from 'react-router-dom';
 import cn from 'classnames';
-import { NavLink } from 'react-router-dom';
 
 type Props = {
-  to: string;
-  title: string;
+  to: To;
+  children: ReactNode;
 };
 
-export const NavigationLink: React.FC<Props> = ({ to, title }) => (
+export const NavigationLink: FC<Props> = ({ to, children }) => (
   <NavLink
     to={to}
     className={({ isActive }) => cn('navbar-item', {
-      'navbar-item has-background-grey-lighter': isActive,
+      'has-background-grey-lighter': isActive,
     })}
   >
-    {title}
+    {children}
   </NavLink>
 );
