+++ src/components/PersonLink/PersonLink.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import cn from 'classnames';
 import { Link } from 'react-router-dom';
+import cn from 'classnames';
 import { Person } from '../../types';
 
 type Props = {
@@ -12,7 +12,7 @@ export const PersonLink: React.FC<Props> = ({ person }) => (
     className={cn({
       'has-text-danger': person.sex === 'f',
     })}
-    to={`../${person.slug}`}
+    to={person.slug}
   >
     {person.name || '-'}
   </Link>
